DB API: * added test code, on our way to full coverage * still some coverage gaps to fill from "newly added DB APIs" * for code coverage, ignore, via /* istanbul ignore next */, caught errors in DB API, as these are somewhat hard to trigger and really are only expected in exceptional cases * for code coverage, ignore, via /* istanbul ignore next */, failure returns attempting to udpate data via Prisma.  These are hard conditions to simulate. * fixed multiple bugs in update() methods dealing with "disconnecting" linked records which are optional.  Sadly, it looks like we need to maintain state that tells us whether or not a linked record ID has been changed.  If so, we need to use the special { disconnect: true } ... and if not, we have to pass in "undefined".  There may be a more elegant approach, but I haven't found it.
The result is some awkward and somewhat fragile code for DB Domain Objects that have nullable foreign key relationships with other Objects.

Given this fragility, I'm circling back with one more update here, so that the framework will require implementing a method to updateCachedValues(), and this method will be called automatically at construction, at create(), and at update(), in order to keep cached values in sync with the use of the object.