Storage:  StorageKey Computation Support In order to compute the storage key, we need to know to which objects the asset is connected.  Until this commit, that was really tough!
The schema has been updated so that Asset has an idSystemObject member.  This represents the *source* object to which the asset belongs.  It's not the same as the SystemObject that represents the Asset.

This schema change causes Prisma to generate some unfortunate and wonky names in the SystemObject type. Instead of "Asset", this is now "Asset_AssetToSystemObject_idAsset".  Yuck.